e2e, libnet: Move cloudinit.go to a sub-package
Currently, KubeVirt supports two cloud-init data sources:
1. NoCloud [1]
2. ConfigDrive [2]

The network tests are using only the `NoCloud` data source.

Move the `NoCloud` code to a dedicated package, in order
to add the `ConfigDrive` code in as a sibling package
in a following commit.

[1] https://cloudinit.readthedocs.io/en/latest/reference/datasources/nocloud.html
[2] https://cloudinit.readthedocs.io/en/latest/reference/datasources/configdrive.html
