+++ tests/BUILD.bazel
@@ -34,7 +34,7 @@ go_library(
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
         "//tests/libdv:go_default_library",
-        "//tests/libnet:go_default_library",
+        "//tests/libnet/cloudinit/nocloud:go_default_library",
         "//tests/libnode:go_default_library",
         "//tests/libstorage:go_default_library",
         "//tests/libvmi:go_default_library",
@@ -226,6 +226,7 @@ go_test(
         "//tests/libinfra:go_default_library",
         "//tests/libinstancetype:go_default_library",
         "//tests/libnet:go_default_library",
+        "//tests/libnet/cloudinit/nocloud:go_default_library",
         "//tests/libnet/job:go_default_library",
         "//tests/libnet/service:go_default_library",
         "//tests/libnode:go_default_library",
