+++ tests/libnet/cloudinit/nocloud/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["cloudinit.go"],
+    importpath = "kubevirt.io/kubevirt/tests/libnet/cloudinit/nocloud",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//tests/libnet:go_default_library",
+        "//vendor/sigs.k8s.io/yaml:go_default_library",
+    ],
+)
