+++ tests/libnet/cloudinit/nocloud/cloudinit.go
@@ -17,12 +17,14 @@
  *
  */
 
-package libnet
+package nocloud
 
 import (
 	"fmt"
 
 	"sigs.k8s.io/yaml"
+
+	"kubevirt.io/kubevirt/tests/libnet"
 )
 
 type NetworkDataOption func(*CloudInitNetworkData) error
@@ -108,13 +110,13 @@ func WithGateway6(gateway6 string) NetworkDataInterfaceOption {
 
 func WithNameserverFromCluster() NetworkDataInterfaceOption {
 	return func(networkDataInterface *CloudInitInterface) error {
-		dnsServerIP, err := ClusterDNSServiceIP()
+		dnsServerIP, err := libnet.ClusterDNSServiceIP()
 		if err != nil {
 			return fmt.Errorf("failed defining network data nameservers when retrieving cluster DNS service IP: %w", err)
 		}
 		networkDataInterface.Nameservers = CloudInitNameservers{
 			Addresses: []string{dnsServerIP},
-			Search:    SearchDomains(),
+			Search:    libnet.SearchDomains(),
 		}
 		return nil
 	}
