+++ tests/network/BUILD.bazel
@@ -55,6 +55,7 @@ go_library(
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
         "//tests/libnet:go_default_library",
+        "//tests/libnet/cloudinit/nocloud:go_default_library",
         "//tests/libnet/cluster:go_default_library",
         "//tests/libnet/job:go_default_library",
         "//tests/libnet/service:go_default_library",
