+++ tests/network/vmi_multus.go
@@ -48,6 +48,7 @@ import (
 	"kubevirt.io/kubevirt/tests"
 	"kubevirt.io/kubevirt/tests/console"
 	"kubevirt.io/kubevirt/tests/libnet"
+	"kubevirt.io/kubevirt/tests/libnet/cloudinit/nocloud"
 	"kubevirt.io/kubevirt/tests/libnode"
 	"kubevirt.io/kubevirt/tests/libvmi"
 	"kubevirt.io/kubevirt/tests/libwait"
@@ -194,10 +195,10 @@ var _ = SIGDescribe("[Serial]Multus", Serial, decorators.Multus, func() {
 			var networkData string
 			BeforeEach(func() {
 				libnet.SkipWhenClusterNotSupportIpv4()
-				networkData, err = libnet.NewNetworkData(
-					libnet.WithEthernet("eth0",
-						libnet.WithDHCP4Enabled(),
-						libnet.WithNameserverFromCluster(),
+				networkData, err = nocloud.NewNetworkData(
+					nocloud.WithEthernet("eth0",
+						nocloud.WithDHCP4Enabled(),
+						nocloud.WithNameserverFromCluster(),
 					),
 				)
 				Expect(err).NotTo(HaveOccurred())
@@ -281,10 +282,10 @@ var _ = SIGDescribe("[Serial]Multus", Serial, decorators.Multus, func() {
 		Context("VirtualMachineInstance with multus network as default network", func() {
 			It("[test_id:1751]should create a virtual machine with one interface with multus default network definition", func() {
 				libnet.SkipWhenClusterNotSupportIpv4()
-				networkData, err := libnet.NewNetworkData(
-					libnet.WithEthernet("eth0",
-						libnet.WithDHCP4Enabled(),
-						libnet.WithNameserverFromCluster(),
+				networkData, err := nocloud.NewNetworkData(
+					nocloud.WithEthernet("eth0",
+						nocloud.WithDHCP4Enabled(),
+						nocloud.WithNameserverFromCluster(),
 					),
 				)
 				Expect(err).NotTo(HaveOccurred())
@@ -811,11 +812,11 @@ func runSafeCommand(vmi *v1.VirtualMachineInstance, command string) error {
 }
 
 func cloudInitNetworkDataWithStaticIPsByMac(nicName, macAddress, ipAddress string) string {
-	networkData, err := libnet.NewNetworkData(
-		libnet.WithEthernet(nicName,
-			libnet.WithAddresses(ipAddress),
-			libnet.WithNameserverFromCluster(),
-			libnet.WithMatchingMAC(macAddress),
+	networkData, err := nocloud.NewNetworkData(
+		nocloud.WithEthernet(nicName,
+			nocloud.WithAddresses(ipAddress),
+			nocloud.WithNameserverFromCluster(),
+			nocloud.WithMatchingMAC(macAddress),
 		),
 	)
 	ExpectWithOffset(1, err).ToNot(HaveOccurred(), "should successfully create static IPs by mac address cloud init network data")
@@ -823,10 +824,10 @@ func cloudInitNetworkDataWithStaticIPsByMac(nicName, macAddress, ipAddress strin
 }
 
 func cloudInitNetworkDataWithStaticIPsByDevice(deviceName, ipAddress string) string {
-	networkData, err := libnet.NewNetworkData(
-		libnet.WithEthernet(deviceName,
-			libnet.WithAddresses(ipAddress),
-			libnet.WithNameserverFromCluster(),
+	networkData, err := nocloud.NewNetworkData(
+		nocloud.WithEthernet(deviceName,
+			nocloud.WithAddresses(ipAddress),
+			nocloud.WithNameserverFromCluster(),
 		),
 	)
 	ExpectWithOffset(1, err).ToNot(HaveOccurred(), "should successfully create static IPs by device name cloud init network data")
